initial files added
To run this face detection and recognition model. First run the face_dataset.py file, it will detect faces and store it in the dataset in the form of jpg images. Then run the training.py, it will train the model to recognize face, and then at last run face_recognition.py to start recognizing faces.